Add "Product" domain and making relationship with "Categoria"

+++ src/main/java/com/marcelo/cursomc/resources/CategoriaResource.java
@@ -18,7 +18,7 @@ public class CategoriaResource {
 	private CategoriaService service;
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<?> find(@PathVariable long id) {
+	public ResponseEntity<?> find(@PathVariable Integer id) {
 		Categoria categoria = service.find(id);
 
 		return ResponseEntity.ok().body(categoria);
